+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Ensure `aws-node` daemonset does not schedule on upgraded nodes.
+- Cleanup `aws-node` resources after a successful migration.
 
 ## [13.0.0-alpha2] - 2022-07-27